+++ util-jvm/src/test/scala/com/twitter/jvm/JvmStatsTest.scala
@@ -14,21 +14,12 @@ class JvmStatsTest extends AnyFunSuite {
     val receiver = new InMemoryStatsReceiver()
     JvmStats.register(receiver)
 
-    val role = Map((ExpressionSchema.Role, "jvm"))
-    val markSweepLabel = role ++ Map(("gc_pool", "PS_MarkSweep"))
-    val scavengeLabel = role ++ Map(("gc_pool", "PS_Scavenge"))
+    val expressions = receiver.getAllExpressionsWithLabel(ExpressionSchema.Role, "jvm")
 
-    val uptimeKey = ExpressionSchemaKey("jvm_uptime", role, Seq())
-    val cyclesKey = ExpressionSchemaKey("gc_cycles", role, Seq())
-    val latencyKey = ExpressionSchemaKey("gc_latency", role, Seq())
+    assert(expressions.size == 7)
 
-    assertMetric(receiver.expressions(uptimeKey))
-    assertMetric(receiver.expressions(cyclesKey))
-    assertMetric(receiver.expressions(latencyKey))
-
-    assertMetric(receiver.expressions(ExpressionSchemaKey("gc_cycles", markSweepLabel, Seq())))
-    assertMetric(receiver.expressions(ExpressionSchemaKey("gc_latency", markSweepLabel, Seq())))
-    assertMetric(receiver.expressions(ExpressionSchemaKey("gc_cycles", scavengeLabel, Seq())))
-    assertMetric(receiver.expressions(ExpressionSchemaKey("gc_latency", scavengeLabel, Seq())))
+    expressions.foreach { x: (ExpressionSchemaKey, ExpressionSchema) =>
+      assertMetric(x._2)
+    }
   }
 }